+++ test/router.js
@@ -1,7 +1,9 @@
 import test from 'ava';
 
 import {
-  createHistory
+  createHistory,
+  syncHistoryWithImmutableStore
 } from 'src/router';
 
 test.todo('createHistory');
+test.todo('syncHistoryWithImmutableStore');